added foreing key checks

+++ back-end/controllers/artista_controller.py
@@ -11,7 +11,10 @@ def get_all_artists():
 
 @router.get("/{artist_id}")
 def get_artist_by_id(artist_id: int):
-  return artista_service.get_artist_by_id(artist_id)
+  artist = artista_service.get_artist_by_id(artist_id)
+  if not artist:
+    raise HTTPException(status_code=404, detail="Artista nao encontrado")
+  return artist
 
 
 @router.post("/")
@@ -21,9 +24,15 @@ def create_artist(artist: ArtistaCreate):
 
 @router.put("/{artist_id}")
 def update_artist(artist_id: int, artist: ArtistaCreate):
+  success = artista_service.update_artist(artist_id, artist)
+  if not success:
+    raise HTTPException(status_code=404, detail="Artista nao encontrado para atualizacao")
   return artista_service.update_artist(artist_id, artist)
 
 
 @router.delete("/{artist_id}")
 def delete_artist(artist_id: int):
+  success = artista_service.delete_artist(artist_id)
+  if not success:
+    raise HTTPException(status_code=404, detail="Artista nao encontrado para exclusao")
   return artista_service.delete_artist(artist_id)
